Fix test regression due to refactoring (moved code around).
